+++ editFile.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>JDATA-Edit</title>
+</head>
+<body>
+
+<script src="/scripts/jquery.min.js"></script>
+<script src="/scripts/common.js"></script>
+</body>
+</html>
