+++ index.html
@@ -4,7 +4,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <title>JDATA</title>
 </head>
 <body>
 
@@ -18,5 +18,6 @@
 
 <script src="/scripts/jquery.min.js"></script>
 <script src="/scripts/common.js"></script>
+<script src="/scripts/index.js"></script>
 </body>
 </html>
